+++ vm/os_fuchsia.cc
@@ -28,7 +28,7 @@ int64_t OS::CurrentMonotonicNanos() {
 const char* OS::Name() { return "fuchsia"; }
 
 
-int OS::NumberOfAvailableProcessors() {
+intptr_t OS::NumberOfAvailableProcessors() {
   return zx_system_get_num_cpus();
 }
